requirements/base.txt
Django>=1.8.3
psycopg2>=2.6.1
django-allauth>=0.23
django-braces>=1.8.1
